Rename workspaces to routes
