+++ packages/graph-explorer/src/routes/DefaultLayout.tsx
