--- packages/graph-explorer/src/workspaces/index.ts
@@ -1,1 +1,0 @@
-export { default as Explorer } from "./GraphExplorer";
